Renamed project to iot-swarm
